requirements.txt
flask==3.0.1
faiss-gpu==1.7.2
opencv-python
Pillow
tqdm
git+https://github.com/openai/CLIP.git
numpy
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
timm
einops
flash_attn
plotly